requirements.txt
asgiref==3.8.1
Django==5.0.4
django-browser-reload==1.18.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
pillow==11.0.0
PyJWT==2.10.1
setuptools==69.2.0
sqlparse==0.5.0
tzdata==2024.1
wheel==0.43.0
